refactor(tests): enhance booking flow tests with improved doctor selection logic
- Updated the booking flow tests to utilize the new pickFirstAvailableBookingDay helper function, allowing for more efficient selection of available booking days.
- Adjusted the URL structure for doctor profile navigation to ensure consistency across environments.
- Improved error handling and messaging when no bookable calendar days are found, enhancing test clarity and reliability.
